cleaning and code coverage
