+++ ParserTests/lexer/BadEmptyStringDelimiter.cs
@@ -1,0 +1,9 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum BadEmptyStringDelimiter
+    {
+        [Lexeme(GenericToken.String, "")] Empty
+    }
+}
