+++ ParserTests/lexer/BadEscapeStringDelimiterLetter.cs
@@ -1,0 +1,9 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum BadEscapeStringDelimiterLetter
+    {
+        [Lexeme(GenericToken.String, "'", "a")] toolong
+    }
+}
