+++ ParserTests/lexer/BadEscapeStringDelimiterTooLong.cs
@@ -1,0 +1,9 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum BadEscapeStringDelimiterTooLong
+    {
+        [Lexeme(GenericToken.String, "'", ";:")] toolong
+    }
+}
