+++ ParserTests/lexer/BadLetterStringDelimiter.cs
@@ -1,0 +1,9 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum BadLetterStringDelimiter
+    {
+        [Lexeme(GenericToken.String, "a")] Letter
+    }
+}
