+++ ParserTests/lexer/CharTokens.cs
@@ -1,0 +1,9 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum CharTokens {
+        [Lexeme(GenericToken.Char,"'","\\")]
+        MyChar
+    }
+}
