+++ ParserTests/lexer/Issue210Token.cs
@@ -1,0 +1,13 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum Issue210Token
+    {
+        EOF = 0,
+
+        [Lexeme(GenericToken.Extension)] SPECIAL,
+        [Lexeme(GenericToken.KeyWord, "x")] X,
+        [Lexeme(GenericToken.SugarToken, "?")] QMARK
+    }
+}
