+++ ParserTests/lexer/ShortExtensions.cs
@@ -1,0 +1,11 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum ShortExtensions
+    {
+        [Extension] DATE,
+
+        [Double] DOUBLE
+    }
+}
