+++ ParserTests/lexer/StringDelimiters.cs
@@ -1,0 +1,9 @@
+using sly.lexer;
+
+namespace ParserTests.lexer
+{
+    public enum StringDelimiters {
+        [Lexeme(GenericToken.String,"'","'")]
+        MyString
+    }
+}
